Добавь GameModel
